RTE-1001: Add qemu based parent image builder.
